# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "local_bridge",
    "description": "Media AI task queue bridge with OpenAPI docs",
    "version": "1.0.0"
  },
  "paths": {
...
    "/v1/single/model-image": {
      "post": {
        "tags": [
          "single",
          "single"
        ],
        "summary": "Create Model Image",
        "operationId": "create_model_image_v1_single_model_image_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModelImageCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelImageCreatedResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response 400 Create Model Image V1 Single Model Image Post"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response 404 Create Model Image V1 Single Model Image Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/single/style-image": {
      "post": {
        "tags": [
          "single",
          "single"
        ],
        "summary": "Create Style Image",
        "operationId": "create_style_image_v1_single_style_image_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StyleImageCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StyleImageCreatedResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response 400 Create Style Image V1 Single Style Image Post"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response 404 Create Style Image V1 Single Style Image Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "GptModelImageJob": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "caseFile": {
            "type": "string",
            "title": "Casefile"
          },
          "mediaAi": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/GptModelImageMediaAi"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "caseFile"
        ],
        "title": "GptModelImageJob"
      },
      "GptModelImageMediaAi": {
        "properties": {
          "kind": {
            "type": "string",
            "title": "Kind",
            "default": "model-image"
          },
          "platform": {
            "type": "string",
            "title": "Platform",
            "default": "gpt"
          },
          "baseUrl": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Baseurl"
          },
          "productId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Productid"
          },
          "productName": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Productname"
          },
          "ipId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ipid"
          },
          "uploadSubDir": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Uploadsubdir"
          }
        },
        "type": "object",
        "title": "GptModelImageMediaAi",
        "description": "mediaAi for /v1/single/model-image."
      },
      "GptStyleImageJob": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "caseFile": {
            "type": "string",
            "title": "Casefile"
          },
          "mediaAi": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/GptStyleImageMediaAi"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "caseFile"
        ],
        "title": "GptStyleImageJob"
      },
      "GptStyleImageMediaAi": {
        "properties": {
          "kind": {
            "type": "string",
            "title": "Kind",
            "default": "style-image"
          },
          "platform": {
            "type": "string",
            "title": "Platform",
            "default": "gpt"
          },
          "baseUrl": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Baseurl"
          },
          "modelImageId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Modelimageid"
          },
          "poseId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Poseid"
          },
          "uploadSubDir": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Uploadsubdir"
          }
        },
        "type": "object",
        "title": "GptStyleImageMediaAi",
        "description": "mediaAi for /v1/single/style-image."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ModelImageCreateRequest": {
        "properties": {
          "modelImageId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Modelimageid"
          },
          "productId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Productid"
          },
          "ipId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ipid"
          },
          "force": {
            "type": "boolean",
            "title": "Force",
            "default": false
          }
        },
        "type": "object",
        "title": "ModelImageCreateRequest"
      },
      "ModelImageCreatedResponse": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok"
          },
          "dryRun": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dryrun"
          },
          "caseFile": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Casefile"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "job": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/GptModelImageJob"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "ok"
        ],
        "title": "ModelImageCreatedResponse"
      },
...
      "StyleImageCreateRequest": {
        "properties": {
          "modelImageId": {
            "type": "string",
            "title": "Modelimageid"
          },
          "poseId": {
            "type": "string",
            "title": "Poseid"
          },
          "force": {
            "type": "boolean",
            "title": "Force",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "modelImageId",
          "poseId"
        ],
        "title": "StyleImageCreateRequest"
      },
      "StyleImageCreatedResponse": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok"
          },
          "dryRun": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dryrun"
          },
          "caseFile": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Casefile"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "job": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/GptStyleImageJob"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "ok"
        ],
        "title": "StyleImageCreatedResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}